# docs/openapi.yaml
openapi: 3.1.0
info:
  title: go-health probe endpoints
  version: 0.1.3
  description: |
    Kubernetes-style health probes exposed by the go-health Probe.
    Field-presence rules are documented per property; they are enforced by
    the golden-file wire-format test, not by OpenAPI keywords.
  license:
    name: MIT
    identifier: MIT
# Health endpoints are unauthenticated by design: the kubelet and load
# balancers poll them without credentials. Auth (if any) is the host's
# reverse-proxy decision, not the probe's.
security: []
# Routes are mounted wherever the host application chooses; all paths are
# relative to that base.
servers:
  - url: /
paths:
  /healthz:
    get:
      summary: Liveness
      description: Always 200 while the process runs and has not started
        shutting down. Performs zero dependency checks; checks is always
        empty. Used by kubelet to decide container restarts.
      operationId: liveness
      responses:
        "200":
          description: The process is alive.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/HealthResponse" }
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
  /readyz:
    get:
      summary: Readiness
      description: 200 when the instance can serve traffic (roll-up pass or
        warn), 503 when any critical service fails or the probe is shutting
        down. Served from the background cache when refresh interval is set.
      operationId: readiness
      responses:
        "200":
          description: Ready (pass) or degraded (warn, non-critical failure).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/HealthResponse" }
        "503":
          description: A critical service failed or the probe is shutting down.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/HealthResponse" }
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
  /startupz:
    get:
      summary: Startup
      description: 503 until every critical service has passed one evaluation,
        then latches to 200 permanently (per process) without re-checking.
      operationId: startup
      responses:
        "200":
          description: Startup complete (latched).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/HealthResponse" }
        "503":
          description: Boot still in progress (latch unset).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/HealthResponse" }
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
components:
  responses:
    MethodNotAllowed:
      description: Non-GET request when method enforcement is enabled
        (WithGETOnly / WithAllowedMethods). The Allow header lists the
        accepted methods, sorted, GET always first.
      headers:
        Allow:
          description: Comma-separated accepted methods, e.g. "GET, HEAD".
          schema: { type: string }
  schemas:
    HealthResponse:
      type: object
      required: [status, shutting_down, total_latency_ms, checks]
      properties:
        status:
          type: string
          enum: [pass, warn, fail]
          description: Roll-up. fail = critical failure or shutdown;
            warn = only non-critical failures; pass = all healthy.
        version:
          type: string
          description: Application version when configured via WithVersion;
            omitted otherwise.
        instance_id:
          type: string
          description: Replica identifier when configured via WithInstanceID;
            omitted otherwise.
        uptime:
          type: string
          description: Human-readable time since boot (Go duration string,
            second resolution); omitted when zero.
        shutting_down:
          type: boolean
          description: Always present. True once Shutdown has been called;
            readiness stays 503 from then on, liveness stays 200.
        total_latency_ms:
          type: integer
          format: int64
          description: Always present. Wall-clock milliseconds of the last
            health-check batch; 0 for liveness.
        timestamp:
          type: string
          format: date-time
          description: RFC 3339 completion time of the evaluation. Absent
            until the first evaluation (omitzero).
        checks:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/Check"
          description: Always present (possibly empty). Keyed by service name.
    Check:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [pass, warn, fail]
        error:
          type: string
          description: Failure message when the check did not pass; omitted
            otherwise.
        since:
          type: string
          format: date-time
          description: RFC 3339 time the probe first observed this check in
            its current status; carried forward unchanged while the status
            holds, restarted on every status change. Absent (omitzero) for
            checks the probe never built (liveness's empty set, the Healthz
            synthetic startup entry). Resets on process restart.
        duration_ns:
          type: integer
          format: int64
          description: Nanoseconds the most recent execution of this check
            took, as reported by its executor. Absent (omitzero) when
            unknown — the raw samber/do injector path cannot measure
            per-check duration, so only detailed sources
            (NewWithDetailedCheck, DetailedHealthRecorder) populate it.